refactor: structure parts of code into _utils
