+++ espresso_machine/build_package/report.py
@@ -8,6 +8,10 @@ Usage:
 
 import sys
 
+import run_examples
+import criteria
+import _utils
+
 
 def _init_attr_report():
     import criteria
@@ -67,7 +71,7 @@ def _collect_additional_attr(all_results, report):
     _additional_attr = {a for a in p_dir if a not in _standard_attr}
     report["additional"].update(_additional_attr)
 
-def raw_compliance_report(problems_to_check=None, pre_build=True):
+def raw_compliance_report(problems_to_check=None):
     """Run all problems and generate a raw compliance report
     
     A typical raw report looks like:
@@ -88,11 +92,9 @@ def raw_compliance_report(problems_to_check=None, pre_build=True):
         }
     }
     """
-    import run_examples
-    import criteria
     report = dict()
-    problems = run_examples.problems_to_run(problems_specified=problems_to_check)
-    results = run_examples.run_problems(problems, pre_build=pre_build)
+    problems = _utils.problems_to_run(problems_to_check)
+    results = run_examples.run_problems(problems, pre_build=_utils.pre_build())
     for res in results:
         _report_for_problem = dict()
         # problem level report
@@ -150,7 +152,7 @@ def _analyse_compliance(new_report):
     _optional_ok = new_report["optional_count"]["error"] == 0
     return _metadata_ok and _required_ok and _optional_ok
 
-def compliance_report(problems_to_check=None, pre_build=True):
+def compliance_report(problems_to_check=None):
     """Generate a readable compliance report based on running raw report
     
     A typical compliance report looks like:
@@ -167,7 +169,7 @@ def compliance_report(problems_to_check=None, pre_build=True):
         }
     }
     """
-    raw_report = raw_compliance_report(problems_to_check, pre_build)
+    raw_report = raw_compliance_report(problems_to_check)
     new_report = dict()
     for prob_name, prob_report in raw_report.items():
         _new_report = dict()
@@ -309,10 +311,10 @@ def pprint_compliance_report(report):
         else:
             print(cformat(bcolors.FAIL, f"\n{prob} is not API-compliant."))
 
-def capability_report(problems_to_check=None, pre_build=True, timeout=1):
+def capability_report(problems_to_check=None, timeout=1):
     # those with TimeoutError will be approximated to be OK
     sys.argv.extend(["--timeout", str(timeout)])
-    _compliance_report = compliance_report(problems_to_check, pre_build)
+    _compliance_report = compliance_report(problems_to_check)
     _capability_report = dict()
     for prob, res in _compliance_report.items():
         _new_report = dict()
