+++ espresso_machine/build_package/run_examples.py
@@ -7,12 +7,9 @@ $ python espresso_machine/build_package/build.py
 
 import sys
 import os
-import warnings
-import pathlib
 import typing
 
 import _utils
-import build
 
 try:
     from espresso.exceptions import InvalidExampleError
@@ -22,37 +19,6 @@ except ModuleNotFoundError as e:
              "level of the project\n  $ pip install ."
     raise e
 
-
-PKG_NAME = "espresso"
-ROOT = str(pathlib.Path(__file__).resolve().parent.parent.parent)
-CONTRIB_FOLDER = ROOT + "/contrib"
-
-
-def _problem_name_to_class(problem_name):   # e.g. "xray_tomography" -> "XrayTomography"
-    return problem_name.title().replace("_", "")
-
-def get_folder_content(folder_name):
-    names = [name for name in os.listdir(folder_name)]
-    paths = [f"{folder_name}/{name}" for name in names]
-    return names, paths
-
-def problems_to_run(problems_specified: typing.Optional[list] = None):
-    all_problems = get_folder_content(CONTRIB_FOLDER)
-    all_problems_zipped = list(zip(*all_problems))
-    all_problems_zipped = [c for c in all_problems_zipped if "." not in c[0]]
-    if problems_specified is None:
-        return all_problems_zipped
-    else:       # filter by specified problems list
-        problems = [c for c in all_problems_zipped if c[0] in problems_specified]
-        problems_not_in_folder = [
-            c for c in problems_specified if c not in all_problems[0]
-        ]
-        if problems_not_in_folder:
-            warnings.warn(
-                "these examples are not detected in 'contrib' folder: " + 
-                ", ".join(problems_not_in_folder)
-            )
-        return problems
 
 class _ProblemModule:
     """get the parent module of a problem class
@@ -66,13 +32,12 @@ class _ProblemModule:
 
     def __enter__(self):
         if self._pre_build:
-            sys.path.insert(1, CONTRIB_FOLDER)
+            sys.path.insert(1, _utils.CONTRIB_FOLDER)
             return __import__(self._problem_name)
         else:
-            return __import__(PKG_NAME)
+            return __import__(_utils.PKG_NAME)
         
     def __exit__(self, exc_type, exc_value, traceback):
-        # print(sys.modules)
         _to_del = set()
         for key in sys.modules.keys():
             if self._problem_name in key:
@@ -111,7 +76,7 @@ prob_properties = [
     ("inv_cov", "inverse_covariance_matrix"),
 ]
 
-@_utils.timeout(seconds=build.args().timeout)
+@_utils.timeout(seconds=_utils.args().timeout)
 def _run_attr(prob_instance, how_to_get, is_method=True):
     if is_method:
         return how_to_get(prob_instance)
@@ -125,6 +90,16 @@ def _get_result(prob_instance, how_to_get, is_method=True):
     except Exception as e:
         return e
 
+def instantiate_example(problem_class, i):
+    try:
+        prob_instance_i = problem_class(i)
+    except Exception as e:
+        if not isinstance(e, InvalidExampleError):
+            prob_instance_i = e
+        else:
+            raise e
+    return prob_instance_i
+
 def collect_methods_outputs(prob_instance_i, all_outputs):
     for (output_name, how_to_get) in prob_methods:
         all_outputs[output_name] = _get_result(prob_instance_i, how_to_get)
@@ -136,14 +111,8 @@ def collect_properties(prob_instance_i, all_outputs):
 def run_example(problem_class, problem_class_str, i) -> dict:
     # prepare
     all_outputs = dict()
-    try:
-        prob_instance_i = problem_class(i)
-    except Exception as e:
-        if not isinstance(e, InvalidExampleError):
-            prob_instance_i = e
-        else:
-            raise e
-    else:  # collect results
+    prob_instance_i = instantiate_example(problem_class, i)
+    if not isinstance(prob_instance_i, Exception):  # collect results
         collect_methods_outputs(prob_instance_i, all_outputs)
         collect_properties(prob_instance_i, all_outputs)
     all_outputs["prob_instance_str"] = f"{problem_class_str}({i})"
@@ -164,9 +133,13 @@ def run_problem(problem_class, problem_class_str) -> typing.Iterator[dict]:
         i += 1
         yield example_res
 
+def run_cmake_if_needed(prob_path, pre_build):
+    if pre_build and "CMakeLists.txt" in os.listdir(prob_path):
+        raise RuntimeError(prob_path)
+
 def run_problems(problems, pre_build):
     for (prob_name, prob_path) in problems:
-        prob_class_str = _problem_name_to_class(prob_name)
+        prob_class_str = _utils.problem_name_to_class(prob_name)
         try:
             with _ProblemModule(pre_build, prob_name) as parent_module:
                 try:
@@ -191,7 +164,7 @@ def run_problems(problems, pre_build):
 
 def main(problems_specified=None):
     _you_want_to_print_something = False
-    problems = problems_to_run(problems_specified)
+    problems = _utils.problems_to_run(problems_specified)
     results = run_problems(problems, pre_build=True)
     for res in results:
         if _you_want_to_print_something: print(res["problem class"])
