+++ espresso_machine/build_package/test_examples.py
@@ -6,13 +6,12 @@
 import pathlib
 import pytest
 
-import run_examples
-import validate
 import report
+import _utils
 
 
 def _pre_build():
-    _args = validate.args()
+    _args = _utils.args()
     return _args.pre or (not _args.pre and not _args.post)
 
 @pytest.fixture
@@ -20,12 +19,11 @@ def pre_build():
     return _pre_build()
 
 def _all_contribs():
-    _args = validate.args()
     pre = _pre_build()
-    problems = run_examples.problems_to_run(_args.contribs)
+    problems = _utils.problems_to_run()
     print("🥃 Running " + ("pre-" if pre else "post-") + "build tests for the following contributions:")
     print("- " + "\n- ".join([c[0] for c in problems]) + "\n")
-    # results = run_examples.run_problems(problems, pre_build=pre)
+    # results = run_examples.run_problems(problems)
     # return results
     return problems
 
@@ -34,7 +32,7 @@ def contrib(request):
     return request.param
 
 def test_contrib(contrib, pre_build):
-    _report = report.compliance_report([contrib[0]], pre_build)
+    _report = report.compliance_report([contrib[0]])
     for _r in _report.values():
         report.pprint_compliance_report(_report)
         if isinstance(_r["api_compliance"], Exception):
